requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
ansi2html==1.9.1
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.21
click==8.1.7
coloredlogs==15.0.1
dash==2.14.2
dash-ag-grid==2.4.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.12.1
dash-table==5.0.0
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.108.0
filelock==3.13.1
finnhub-python==2.4.19
Flask==3.0.0
flatbuffers==23.5.26
frozendict==2.3.10
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.25.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jq==1.6.0
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
lxml==5.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
peewee==3.17.0
plotly==5.18.0
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.33
zipp==3.17.
